Renamed filters to pipes
The concept of filters in angular 1 is now 'pipes'. Renaming the folder to reflect that. For now I'll leave the module and localizeStringDates filter, since these are still angular 1. For backwards compatibility, we'll still export the module as 'filters' from the main module, in addition to pipes.
