+++ source/pipes/date/date.pipe.md
